web_app/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.12.0
Pillow>=9.5.0
numpy>=1.23.0
streamlit>=1.28.0
